+++ R/hooks.R
@@ -1,3 +1,9 @@
+knitr_opts_current <- function(x, default = FALSE){
+  x <- knitr::opts_current$get(x)
+  if(is.null(x)) x <- default
+  x
+}
+
 #' @importFrom knitr opts_chunk
 #' @title knitr hook for figure caption autonumbering
 #' @description The function allows you to add a hook when executing
@@ -12,13 +18,20 @@ plot_word_fig_caption <- function(x, options) {
     fig.id <- options$label
   else fig.id <- options$fig.id
   if(!is.logical(options$fig.topcaption)) options$fig.topcaption <- FALSE
-  
+
+  tnd <- knitr_opts_current("fig.cap.tnd", default = 0)
+  tns <- knitr_opts_current("fig.cap.tns", default = "-")
+  fig.fp_text <- knitr_opts_current("fig.fp_text", default = fp_text_lite(bold = TRUE))
+
+
   bc <- block_caption(label =  options$fig.cap, style = options$fig.cap.style,
                       autonum = run_autonum(
                         seq_id = gsub(":$", "", options$fig.lp),
                         pre_label = options$fig.cap.pre,
                         post_label = options$fig.cap.sep,
-                        bkm = fig.id, bkm_all = FALSE
+                        bkm = fig.id, bkm_all = FALSE,
+                        tnd = tnd, tns = tns,
+                        prop = fig.fp_text
                       ))
   cap_str <- to_wml(bc, knitting = TRUE)
 
@@ -41,11 +54,12 @@ plot_word_fig_caption <- function(x, options) {
     fig.style_id <- style_id("Normal", type = "paragraph", si)
 
   }
-  ooxml <- paste0("<w:p><w:pPr><w:jc w:val=\"%s\"/><w:pStyle w:val=\"%s\"/></w:pPr>",
+  ooxml <- "<w:p><w:pPr><w:jc w:val=\"%s\"/><w:pStyle w:val=\"%s\"/></w:pPr>"
+  ooxml <- sprintf(ooxml, opts_current$get("fig.align"), fig.style_id)
+  ooxml <- paste0(ooxml,
          to_wml(img),
          "</w:p>"
          )
-  ooxml <- sprintf(ooxml, opts_current$get("fig.align"), fig.style_id)
   img_wml <- paste("```{=openxml}", ooxml, "```", sep = "\n")
 
   if (options$fig.topcaption)