+++ R/rdocx_post_knit.R
@@ -1,6 +1,6 @@
-#' @importFrom officer run_reference to_wml
+#' @importFrom officer run_reference to_wml run_word_field
 as_reference_caption <- function(z){
-  sf <- run_seqfield(seqfield = paste0(" REF ", z, " \\h"))
+  sf <- run_word_field(field = paste0(" REF ", z, " \\h"))
   str <- to_wml(sf)
   paste0("`<w:hyperlink w:anchor=\"", z, "\">", str, "</w:hyperlink>`{=openxml}")
 }
@@ -9,22 +9,28 @@ as_reference_standard <- function(z, numbered = FALSE){
   if(numbered)
     str <- paste0(str, " \\r")
 
-  sf <- run_seqfield(seqfield = str)
+  sf <- run_word_field(field = str)
   str <- to_wml(sf)
   paste0("`<w:hyperlink w:anchor=\"", z, "\">", str, "</w:hyperlink>`{=openxml}")
 }
 
-post_knit_table_captions <- function(content, tab.cap.pre, tab.cap.sep, style) {
-  is_captions <- grepl("<caption>\\(\\\\#tab:[-[:alnum:]]+\\)(.*)</caption>", content)
+post_knit_table_captions <- function(content, tab.cap.pre, tab.cap.sep, style, tab.lp, tnd, tns,
+                                     prop = NULL) {
+  pattern1 <- paste0("\\(\\\\#", tab.lp, "[-[:alnum:]]+\\)(.*)")
+  pattern2 <- paste0("\\(\\\\#", tab.lp, "([-[:alnum:]]+)\\)(.*)")
+
+  is_captions <- grepl(pattern1, content)
   if (any(is_captions)) {
     captions <- content[is_captions]
-    ids <- gsub("<caption>\\(\\\\#tab:([-[:alnum:]]+)\\)(.*)</caption>", "\\1", captions)
-    labels <- gsub("<caption>\\(\\\\#tab:[-[:alnum:]]+\\)(.*)</caption>", "\\1", captions)
+    ids <- gsub(pattern2, "\\1", captions)
+    labels <- gsub(pattern1, "\\1", captions)
     str <- mapply(function(label, id) {
       pandoc_wml_caption(
         cap = label, cap.style = style,
         cap.pre = tab.cap.pre, cap.sep = tab.cap.sep,
-        id = id, seq_id = "tab:"
+        id = id, seq_id = tab.lp,
+        tnd = tnd, tns = tns,
+        prop = prop
       )
     }, label = labels, id = ids, SIMPLIFY = FALSE)
     content[is_captions] <- unlist(str)
@@ -34,8 +40,8 @@ post_knit_table_captions <- function(content, tab.cap.pre, tab.cap.sep, style) {
 
 post_knit_caption_references <- function(content, lp = ""){
 
-  if(!lp %in% c("tab:", "fig:")){
-    stop("lp must be one of `tab:`, `fig:`.")
+  if(!grepl(":$", lp)){
+    stop("lp must end with `:`")
   }
   regexpr_str <- paste0('\\\\@ref\\(', lp, '([-[:alnum:]]+)\\)')